Ajuste no repositorio para salvar exception quando estou dentro do proprio metodo, ajuste em controller, dominio, repositorio e serviço.

+++ WebApi.Ecommerce/Domain/DTOs/BootstrapTablePaginationDTO.cs
@@ -5,6 +5,7 @@ namespace WebApi.Ecommerce.Domain.DTOs
     public class BootstrapTablePaginationDTO<PaginatedEntity>
     {
         public List<PaginatedEntity> Rows { get; set; }
-        public long Total { get; set; }
+
+        public int Total { get; set; }
     }
 }
